Change all instances of long to int32_t Between x86 and arm64 it looks like long is the only type (othe than pointers) that differ in size.

+++ linuxdoom-1.10/d_net.h
@@ -79,7 +79,7 @@ typedef struct
 typedef struct
 {
     // Supposed to be DOOMCOM_ID?
-    long		id;
+    int32_t		id;
     
     // DOOM executes an int to execute commands.
     short		intnum;
